Prisma error handling improved

+++ src/app/api/users/[id]/route.ts
@@ -1,13 +1,20 @@
 import { PrismaClient, Prisma } from "@/generated/prisma";
-import { ErrorI } from "@/types/api";
+import { handlePrismaError } from "@/utils/prisma-error";
 import { NextResponse } from "next/server";
 
 const prisma = new PrismaClient();
 
-export async function DELETE({ params }: { params: Promise<{ id: string }> }) {
-  const id = (await params).id;
-
+export async function DELETE({ params }: { params: { id: string } }) {
   try {
+    const id = Number(params?.id);
+
+    if (isNaN(id)) {
+      return NextResponse.json(
+        { error: "Insert a number as id" },
+        { status: 500 }
+      );
+    }
+
     await prisma.user.delete({ where: { id: Number(id) } });
 
     return NextResponse.json(
@@ -16,31 +23,8 @@ export async function DELETE({ params }: { params: Promise<{ id: string }> }) {
         status: 200,
       }
     );
-  } catch (error) {
-    let errorBody: ErrorI;
-    if (error instanceof Prisma.PrismaClientKnownRequestError) {
-      if (error.code === "P2025") {
-        errorBody = {
-          status: 404,
-          error: "User not found",
-        };
-
-        return NextResponse.json(errorBody, { status: 404 });
-      }
-
-      errorBody = {
-        status: 500,
-        error: error.message,
-      };
-
-      return NextResponse.json(errorBody, { status: 500 });
-    }
-
-    errorBody = {
-      status: 500,
-      error: "Internal server error on Delete",
-    };
-
-    return NextResponse.json(errorBody, { status: 500 });
+  } catch (e) {
+    const { error, status } = handlePrismaError(e, "Delete User");
+    return NextResponse.json({ error }, { status });
   }
 }
